Structure Program Assignment
A Small company/firm  has 10 employees. They are given a pay of 20000 each. Company has decided to give bonus to their employees for February and may month according to their attendance.. Consider following data. 
If attendance >95% bonus is 40% of their salary. 
If attendance >85% bonus is 30% of their salary. 
If attendance >75% bonus is 20% of their salary. 
If attendance >60% bonus is 10% of their salary.
Other will get a bonus of 5% of their salary. 
Employee having below 40% attendance are not eligible for bonus.. 
According to given data find out annual salary, bonus of each employee. Also find out how much amount  company is needed to pay as bonus. Also calculate % bonus of company... 
Write this program and be user friendly.